refactor: Clean up dagster types to reduce mixing of concerns
By having too many types for disparate use cases it encourages mixing of concerns. This
splits the types into more bounded domains.
