+++ src/ol_orchestrate/lib/dagster_types/files.py
@@ -1,0 +1,8 @@
+from pathlib import PosixPath
+
+from dagster import usable_as_dagster_type
+
+
+@usable_as_dagster_type
+class DagsterPath(PosixPath):
+    pass
