+++ src/ol_orchestrate/lib/dagster_types/google.py
@@ -1,0 +1,4 @@
+from dagster import PythonObjectDagsterType
+from google.cloud.bigquery.dataset import DatasetListItem
+
+DatasetDagsterType = PythonObjectDagsterType(DatasetListItem, name="DatasetDagsterType")
